fix: allow for out of order card creation
Previously, we assumed that the `Card` is always added before the
`cardContent` and `cardDeleted`.

This assumption is not true. When we grade the card, the sequence number
of the card will increase beyond the sequence number of the cardContent
and cardDeleted.

Thus, we must be able to handle the out of order creation of cards by
filling in with a default card first.

As long as we have all the operations applied, the card will be in a
valid created state