refactor: shift seqNo into array operation
There is no need to read the `seqNo` each time we execute an operation
as the operations are idempotent.

We can read the seq number once, apply the operations and update it
based on the maximum that we've seen so far.

This makes the application of the operations about 2x faster.