fix: resolve clicking after exiting dialog
There's a known issue where the page becomes unresponsive after
closing the modal dialog.

Specifically, in my case it's being triggered when you right click the
context menu, and then trigger the modal through an action like 'edit'.

I followed this solution:
https://github.com/shadcn-ui/ui/issues/1912
on github which resolves the issue by adding `modal={false}` to the
context menu component.

If this is a persistent issue in the future, perhaps I could just
add `modal={false}` to the dropdown and context menu files.