fix: use explicit variables rather than spread
We were spreading the other card metadata into the card operation.

It doesn't affect the behaviour, but it does use more storage.

In addition, there is a bug where we used crypto.randomUuid() for the
id, when instead we should be using the existing card ID.
The bug was not noticed because the spread operator overrides the
existing cardId.